Put a cap to the maximum number of rows per table at 2bln. Indexes accordingly.

+++ src/deveeldb/Deveel.Data.Sql/ITable.cs
@@ -45,7 +45,7 @@ namespace Deveel.Data.Sql {
 		/// <summary>
 		/// Gets the total number of rows in the table.
 		/// </summary>
-		long RowCount { get; }
+		int RowCount { get; }
 
 		/// <summary>
 		/// Gets a single cell within the table that is
